minor: change types to fix some compiler warnings

+++ tiny_cnn/io/display.h
@@ -33,7 +33,7 @@ class timer
 class progress_display
 {
  public:
-  explicit progress_display( unsigned long expected_count_,
+  explicit progress_display( size_t expected_count_,
                              std::ostream & os = std::cout,
                              const std::string & s1 = "\n", //leading strings
                              const std::string & s2 = "",
@@ -41,7 +41,7 @@ class progress_display
    // os is hint; implementation may ignore, particularly in embedded systems
    :  m_os(os), m_s1(s1), m_s2(s2), m_s3(s3) { restart(expected_count_); }
 
-  void           restart( unsigned long expected_count_ )
+  void           restart( size_t expected_count_ )
   //  Effects: display appropriate scale
   //  Postconditions: count()==0, expected_count()==expected_count_
   {
@@ -55,7 +55,7 @@ class progress_display
     if ( !_expected_count ) _expected_count = 1;  // prevent divide by zero
   } // restart
 
-  unsigned long  operator+=( unsigned long increment )
+  size_t  operator+=( size_t increment )
   //  Effects: Display appropriate progress tic if needed.
   //  Postconditions: count()== original count() + increment
   //  Returns: count().
@@ -64,9 +64,9 @@ class progress_display
     return _count;
   }
 
-  unsigned long  operator++()           { return operator+=( 1 ); }
-  unsigned long  count() const          { return _count; }
-  unsigned long  expected_count() const { return _expected_count; }
+  size_t  operator++()           { return operator+=( 1 ); }
+  size_t  count() const          { return _count; }
+  size_t  expected_count() const { return _expected_count; }
 
   private:
   std::ostream &     m_os;  // may not be present in all imps
@@ -74,23 +74,23 @@ class progress_display
   const std::string  m_s2;  //  const char *, and efficiency or size are
   const std::string  m_s3;  //  not issues
 
-  unsigned long _count, _expected_count, _next_tic_count;
-  unsigned int  _tic;
+  size_t  _count, _expected_count, _next_tic_count;
+  size_t  _tic;
   void display_tic()
   {
     // use of floating point ensures that both large and small counts
     // work correctly.  static_cast<>() is also used several places
     // to suppress spurious compiler warnings. 
-    unsigned int tics_needed =
-      static_cast<unsigned int>(
+    size_t tics_needed =
+      static_cast<size_t>(
         (static_cast<double>(_count)/_expected_count)*50.0 );
     do { m_os << '*' << std::flush; } while ( ++_tic < tics_needed );
     _next_tic_count = 
-      static_cast<unsigned long>((_tic/50.0)*_expected_count);
+      static_cast<size_t>((_tic/50.0)*_expected_count);
     if ( _count == _expected_count ) {
       if ( _tic < 51 ) m_os << '*';
       m_os << std::endl;
-      }
+    }
   } // display_tic
 
   progress_display &operator = (const progress_display &) = delete;
